[pkg/checker] refactored fixmynetboot in pkg/checker and cli
