+++ pkg/checker/colours.go
@@ -1,4 +1,4 @@
-package main
+package checker
 
 import "fmt"
