+++ pkg/checker/dmesg.go
@@ -1,4 +1,4 @@
-package main
+package checker
 
 import (
 	"strings"
